third_party/WebKit/LayoutTests/inspector/console/console-linkify-message-location-expected.txt

CONSOLE MESSAGE: line 8: 239
Test that console.log() would linkify its location in respect with blackboxing.

foo.js:4 
boo.js:4 
VM:1 
foo.js:4 
